% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetEmbeddingLayer.R
\name{getEmbeddingLayer}
\alias{getEmbeddingLayer}
\title{Extract layer-0 embedding from Prov-GigaPath CSV}
\usage{
getEmbeddingLayer(pg_csv, layer = 0)
}
\arguments{
\item{pg_csv}{Path to a Prov-GigaPath CSV file}

\item{layer}{The embedding layer to extract (default 0)}
}
\value{
Numeric vector of embeddings
}
\description{
Extract layer-0 embedding from Prov-GigaPath CSV
}
